# Improve check performance by filtering it's input before parsing