--- a/datadog_checks_base/datadog_checks/checks/prometheus/mixins.py
+++ b/datadog_checks_base/datadog_checks/checks/prometheus/mixins.py
@@ -150,6 +150,11 @@ class PrometheusScraperMixin(object):
         # Timeout used during the network request
         self.prometheus_timeout = 10
 
+        # List of strings to filter the input text payload on. If any line contains
+        # one of these strings, it will be filtered out before being parsed.
+        # INTERNAL FEATURE, might be removed in future versions
+        self._text_filter_blacklist = []
+
     def parse_metric_family(self, response):
         """
         Parse the MetricFamily from a valid requests.Response object to provide a MetricFamily object (see [0])
@@ -188,12 +193,16 @@ class PrometheusScraperMixin(object):
                 yield message
 
         elif 'text/plain' in response.headers['Content-Type']:
+            input_gen = response.iter_lines(chunk_size=self.REQUESTS_CHUNK_SIZE)
+            if self._text_filter_blacklist:
+                input_gen = self._text_filter_input(input_gen)
+
             messages = defaultdict(list)  # map with the name of the element (before the labels)
             # and the list of occurrences with labels and values
 
             obj_map = {}  # map of the types of each metrics
             obj_help = {}  # help for the metrics
-            for metric in text_fd_to_metric_families(response.iter_lines(chunk_size=self.REQUESTS_CHUNK_SIZE)):
+            for metric in text_fd_to_metric_families(input_gen):
                 metric.name = self.remove_metric_prefix(metric.name)
                 metric_name = "%s_bucket" % metric.name if metric.type == "histogram" else metric.name
                 metric_type = self.type_overrides.get(metric_name, metric.type)
@@ -216,6 +225,22 @@ class PrometheusScraperMixin(object):
         else:
             raise UnknownFormatError('Unsupported content-type provided: {}'.format(
                 response.headers['Content-Type']))
+
+    def _text_filter_input(self, input_gen):
+        """
+        Filters out the text input line by line to avoid parsing and processing
+        metrics we know we don't want to process. This only works on `text/plain`
+        payloads, and is an INTERNAL FEATURE implemented for the kubelet check
+        :param input_get: line generator
+        :output: generator of filtered lines
+        """
+        for line in input_gen:
+            for item in self._text_filter_blacklist:
+                if item in line:
+                    break
+            else:
+                # No blacklist matches, passing the line through
+                yield line
 
     def remove_metric_prefix(self, metric):
         return metric[len(self.prometheus_metrics_prefix):] if metric.startswith(self.prometheus_metrics_prefix) else metric